:hammer: Split server code into multiple files
