+++ Pipfile
@@ -14,4 +14,4 @@ pytweening = "*"
 numpy = "*"
 
 [requires]
-python_version = "3.5"
+python_version = "3.7"
